[HumanBenchmark][Backend][Events] Refactor Events backend further
